+++ crates/noseyparker/src/rules_database.rs
@@ -1,42 +1,28 @@
 use anyhow::{bail, Result};
 use vectorscan::{BlockDatabase, Pattern, Flag};
 use regex::bytes::Regex;
-use std::path::Path;
 use std::time::Instant;
 use tracing::{debug, debug_span};
 
-use noseyparker_rules::{Rule, Rules};
-
-use crate::defaults::get_default_rules;
+use noseyparker_rules::Rule;
 
 pub struct RulesDatabase {
     // NOTE: pub(crate) here so that `Matcher` can access these
-    pub(crate) rules: Rules,
+    pub(crate) rules: Vec<Rule>,
     pub(crate) anchored_regexes: Vec<Regex>,
     pub(crate) vsdb: BlockDatabase,
 }
 
 impl RulesDatabase {
-    /// Create a new `RulesDatabase` with the built-in default set of rules.
-    pub fn from_default_rules() -> Result<Self> {
-        Self::from_rules(get_default_rules()?)
-    }
-
-    /// Create a new `RulesDatabase` from rules files found within the given directory.
-    pub fn from_directory<P: AsRef<Path>>(path: P) -> Result<Self> {
-        Self::from_rules(Rules::from_directory(path)?)
-    }
-
-    /// Create a new `RulesDatabase` from the given set of rules.
-    pub fn from_rules(rules: Rules) -> Result<Self> {
+    /// Create a new `RulesDatabase` from the given collection of rules.
+    pub fn from_rules(rules: Vec<Rule>) -> Result<Self> {
         let _span = debug_span!("RulesDatabase::from_rules").entered();
 
-        if rules.rules.is_empty() {
+        if rules.is_empty() {
             bail!("No rules to compile");
         }
 
         let patterns = rules
-            .rules
             .iter()
             .enumerate()
             .map(|(id, r)| {
@@ -51,13 +37,12 @@ impl RulesDatabase {
 
         let t2 = Instant::now();
         let anchored_regexes = rules
-            .rules
             .iter()
             .map(Rule::as_anchored_regex)
             .collect::<Result<Vec<Regex>>>()?;
         let d2 = t2.elapsed().as_secs_f64();
 
-        debug!("Compiled {} rules: vectorscan {}s; regex {}s", rules.rules.len(), d1, d2);
+        debug!("Compiled {} rules: vectorscan {}s; regex {}s", rules.len(), d1, d2);
         Ok(RulesDatabase {
             rules,
             vsdb,
@@ -72,12 +57,14 @@ impl RulesDatabase {
     //     // Ok(())
     // }
 
+    #[inline]
     pub fn num_rules(&self) -> usize {
         self.rules.len()
     }
 
+    #[inline]
     pub fn get_rule(&self, index: usize) -> Option<&Rule> {
-        self.rules.rules.get(index)
+        self.rules.get(index)
     }
 }